refactor: small improvements

+++ src/api/providers/pg-db/migrations/001-customers-table.sql
@@ -4,8 +4,8 @@ CREATE TABLE IF NOT EXISTS customers (
   "doc" TEXT NOT NULL UNIQUE,
   "email" TEXT NOT NULL UNIQUE,
   "password" VARCHAR NOT NULL,
-  "account" BIGINT NOT NULL,
-  "wallet" FLOAT NOT NULL,
+  "account" INT NOT NULL,
+  "wallet" NUMERIC(10, 2) NOT NULL,
   "type" TEXT NOT NULL,
   "created_at" TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 )
