+++ src/api/providers/pg-db/migrations/002-account-sequence.sql
@@ -1,5 +1,5 @@
 CREATE SEQUENCE IF NOT EXISTS account_number
-AS BIGINT
+AS INT
 START WITH 1001
 OWNED BY  customers.account
